LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-no-cr-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-no-cr' failed: Status line does not end with CRLF
Handshake should fail when the first line does not end with CRLF.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS connected is false
PASS origin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
